bot/requirements.txt
requests~=2.22.0
python-dotenv==0.18.0